Refactorisation: reorganiser les tests d'intégration et améliorer la structure des composants
