+++ back/src/test/java/com/openclassrooms/mddapi/integration/ThemeControllerTest.java
@@ -1,5 +1,6 @@
-package com.openclassrooms.mddapi;
+package com.openclassrooms.mddapi.integration;
 
+import com.openclassrooms.mddapi.config.SecurityConfig;
 import com.openclassrooms.mddapi.controller.ThemeController;
 import com.openclassrooms.mddapi.dto.ThemeResponse;
 import com.openclassrooms.mddapi.security.JwtTokenProvider;
@@ -9,6 +10,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.context.annotation.Import;
 import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.web.servlet.MockMvc;
 
@@ -21,16 +23,16 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 /**
- * Tests d'intégration — couche HTTP de ThemeController.
+ * Tests de la couche HTTP — ThemeController.
  * Vérifie l'accès sans auth (401), la liste des thèmes et l'abonnement.
  */
 @WebMvcTest(ThemeController.class)
-class ThemeControllerIT {
+@Import(SecurityConfig.class)
+class ThemeControllerTest {
 
     @Autowired private MockMvc mockMvc;
 
     @MockBean private ThemeService themeService;
-    // Beans requis par SecurityConfig (JwtAuthenticationFilter) — non utilisés directement dans les tests
     @MockBean private UserDetailsServiceImpl userDetailsService;
     @MockBean private JwtTokenProvider jwtTokenProvider;
 
@@ -68,9 +70,7 @@ class ThemeControllerIT {
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.message").value("Abonnement au thème effectué avec succès"));
 
-        // Vérifie que le service a bien été invoqué
         verify(themeService).subscribe(1L);
     }
 }
-
