+++ front/src/app/pages/profile/profile.component.scss
@@ -130,45 +130,47 @@ mat-card-actions {
 }
 
 .subscriptions-list {
-  display: flex;
-  flex-direction: column;
+  display: grid;
+  grid-template-columns: 1fr 1fr;
   gap: 12px;
+
+  @media (max-width: 599px) {
+    grid-template-columns: 1fr;
+  }
 }
 
 .subscription-item {
-  padding: 0;
-
-  .subscription-content {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    gap: 16px;
-    padding: 12px 16px;
-  }
-
-  .subscription-info {
-    display: flex;
-    flex-direction: column;
-    flex: 1;
-    min-width: 0;
-  }
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  text-align: center;
+  background: #f5f5f5;
+  border-radius: 8px;
+  padding: 20px 16px;
+  gap: 8px;
 
   .subscription-title {
-    font-size: 15px;
-    font-weight: 600;
-    margin-bottom: 2px;
+    font-size: 16px;
+    font-weight: 700;
+    color: rgba(0, 0, 0, 0.87);
+    overflow-wrap: break-word;
+    word-break: break-word;
   }
 
   .subscription-description {
     font-size: 13px;
     color: rgba(0, 0, 0, 0.6);
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
+    white-space: normal;
+    overflow-wrap: break-word;
+    word-break: break-word;
+    margin-bottom: 8px;
   }
 
   .unsubscribe-btn {
-    flex-shrink: 0;
     font-size: 13px;
+    width: 100%;
   }
 }
+
+
+
