+++ front/src/app/pages/themes/themes.component.html
@@ -39,7 +39,7 @@
       <mat-card-actions>
         <button
           mat-raised-button
-          [color]="theme.subscribed ? undefined : 'primary'"
+          color="primary"
           [disabled]="theme.subscribed"
           (click)="subscribe(theme)"
         >
