+++ front/src/styles/styles.scss
@@ -5,6 +5,7 @@ body { margin: 0; font-family: Roboto, "Helvetica Neue", sans-serif; }
 
 mat-card, .mat-card {
   box-shadow: none !important;
+  padding: 0 !important;
 }
 
 @media (max-width: 599px) {
